plcontrol/lantern/requirements.txt
zmq
byt
ruamel.yaml
pyserial
